Made implicit conversion from size_t to int explicit (silence warning).

+++ mkd2html.c
@@ -161,7 +161,7 @@ char **argv;
 
     if ( h ) {
 	fprintf(output,"  <title>");
-	mkd_generateline(h, strlen(h), output, 0);
+	mkd_generateline(h, (int)strlen(h), output, 0);
 	fprintf(output, "</title>\n");
     }
     for ( i=0; i < S(headers); i++ )
